last page and next page.

+++ server/Assets/Scripts/keyboard.cs
@@ -7,23 +7,58 @@ public class Keyboard : MonoBehaviour {
     public Text screen;
     private RectTransform hoverKey = null;
     private ArrayList wordList;
+    private int selectNum = 0;
+    private int page = 0;
 
     // Use this for initialization
     void Start () {
-
+        calnSelectNum();
     }
 	
 	// Update is called once per frame
 	void Update () {
+        updateHover();
+        updatePageButton();
+    }
+
+    void calnSelectNum() {
+        foreach (RectTransform key in transform) {
+            if (key.tag == "select") {
+                selectNum++;
+            }
+        }
+    }
+
+    void updateHover() {
         hoverKey = null;
         foreach (RectTransform key in transform) {
             if (cursorInsideKey(key)) {
                 hoverKey = key;
                 setKeyColor(key, Color.yellow);
-            } else {
+            }
+            else {
                 setKeyColor(key, Color.white);
             }
         }
+    }
+
+    void updatePageButton() {
+        foreach (RectTransform key in transform) {
+            if (key.name == "lastPage") {
+                key.GetComponentInChildren<RawImage>().enabled = canLastPage();
+            }
+            if (key.name == "nextPage") {
+                key.GetComponentInChildren<RawImage>().enabled = canNextPage();
+            }
+        }
+    }
+
+    bool canLastPage() {
+        return page - 1 >= 0;
+    }
+
+    bool canNextPage() {
+        return page + 1 < (wordList.Count + selectNum - 1) / selectNum;
     }
 
     void setKeyColor(RectTransform key, Color color) {
@@ -48,34 +83,64 @@ public class Keyboard : MonoBehaviour {
     }
 
     public void confirm() {
-        if (hoverKey != null) {
-            if (hoverKey.name == "keyDelete") {
-                screen.text = deleteWord(screen.text);
-                GetComponent<Dictionary>().clearPos();
-                return;
+        if (hoverKey != null && hoverKey.tag == "delete") {
+            screen.text = deleteWord(screen.text);
+            page = 0;
+            GetComponent<Dictionary>().clearPos();
+            return;
+        }
+
+        if (hoverKey != null && hoverKey.tag == "select") {
+            screen.text = deleteWord(screen.text) + hoverKey.GetComponentInChildren<Text>().text;
+            page = 0;
+            GetComponent<Dictionary>().clearPos();
+            clearSelect();
+            return;
+        }
+        
+        if (hoverKey != null && hoverKey.tag == "page") {
+            if (hoverKey.name == "lastPage" && canLastPage()) {
+                page--;
             }
-            if (hoverKey.name.Length >= 7 && hoverKey.name.Substring(0, 6) == "select") {
-                screen.text = deleteWord(screen.text) + hoverKey.GetComponentInChildren<Text>().text;
-                GetComponent<Dictionary>().clearPos();
-                return;
+            if (hoverKey.name == "nextPage" && canNextPage()) {
+                page++;
             }
+            drawSelect();
+            return;
         }
 
         wordList = GetComponent<Dictionary>().getWordList();
+        GetComponent<Dictionary>().clearPos();
+
+        drawSelect();
+        drawDefaultWord();
+    }
+
+    void drawSelect() {
         foreach (RectTransform key in transform) {
-            if (key.name.Length >= 7 && key.name.Substring(0, 6) == "select") {
-                int rank = key.name[6] - '0';
+            if (key.tag == "select") {
+                int rank = page * selectNum + (key.name[6] - '0');
                 if (rank < wordList.Count) {
                     key.GetComponentInChildren<Text>().text = ((Dictionary.Word)wordList[rank]).word + " ";
-                }
-                else {
+                } else {
                     key.GetComponentInChildren<Text>().text = "";
                 }
             }
         }
+    }
+
+    void clearSelect() {
+        foreach (RectTransform key in transform) {
+            if (key.tag == "select") {
+                key.GetComponentInChildren<Text>().text = "";
+            }
+        }
+    }
+
+    void drawDefaultWord() {
         if (wordList.Count > 0) {
             screen.text += ((Dictionary.Word)wordList[0]).word + " ";
-            GetComponent<Dictionary>().clearPos();
+            page = 0;
         }
     }
 }
